[#749] Move common_checks functions to lib directories
