+++ cove/views.py
@@ -1,6 +1,4 @@
-import json
 import logging
-import os
 from datetime import timedelta
 
 from django.db.models.aggregates import Count
@@ -8,7 +6,6 @@ from django.shortcuts import render
 from django.utils.translation import ugettext_lazy as _
 from django.utils import timezone
 
-import cove.lib.common as common
 from cove.input.models import SuppliedData
 from cove.lib.tools import get_file_type
 
@@ -54,73 +51,6 @@ def explore_data_context(request, pk):
     }
 
     return (context, data, None)
-
-
-def common_checks_context(upload_dir, json_data, schema_obj, schema_name, context, extra_checkers=None, fields_regex=False, api=False):
-    schema_version = getattr(schema_obj, 'version', None)
-    schema_version_choices = getattr(schema_obj, 'version_choices', None)
-
-    if schema_version:
-        schema_version_display_choices = tuple(
-            (version, display_url[0]) for version, display_url in schema_version_choices.items()
-        )
-        context['version_used'] = schema_version
-        if not api:
-            context.update({
-                'version_display_choices': schema_version_display_choices,
-                'version_used_display': schema_version_choices[schema_version][0]}
-            )
-
-    additional_fields = sorted(common.get_counts_additional_fields(json_data, schema_obj, schema_name,
-                                                                   context, fields_regex=fields_regex))
-    context.update({
-        'data_only': additional_fields,
-        'additional_fields_count': sum(item[2] for item in additional_fields)
-    })
-
-    cell_source_map = {}
-    heading_source_map = {}
-    if context['file_type'] != 'json':  # Assume it is csv or xlsx
-        with open(os.path.join(upload_dir, 'cell_source_map.json')) as cell_source_map_fp:
-            cell_source_map = json.load(cell_source_map_fp)
-
-        with open(os.path.join(upload_dir, 'heading_source_map.json')) as heading_source_map_fp:
-            heading_source_map = json.load(heading_source_map_fp)
-
-    validation_errors_path = os.path.join(upload_dir, 'validation_errors-2.json')
-    if os.path.exists(validation_errors_path):
-        with open(validation_errors_path) as validiation_error_fp:
-            validation_errors = json.load(validiation_error_fp)
-    else:
-        validation_errors = common.get_schema_validation_errors(json_data, schema_obj, schema_name,
-                                                                cell_source_map, heading_source_map,
-                                                                extra_checkers=extra_checkers)
-        with open(validation_errors_path, 'w+') as validiation_error_fp:
-            validiation_error_fp.write(json.dumps(validation_errors))
-
-    extensions = None
-    if getattr(schema_obj, 'extensions', None):
-        extensions = {
-            'extensions': schema_obj.extensions,
-            'invalid_extension': schema_obj.invalid_extension,
-            'is_extended_schema': schema_obj.extended,
-            'extended_schema_url': schema_obj.extended_schema_url
-        }
-
-    context.update({
-        'schema_url': schema_obj.release_pkg_schema_url,
-        'extensions': extensions,
-        'validation_errors': sorted(validation_errors.items()),
-        'validation_errors_count': sum(len(value) for value in validation_errors.values()),
-        'deprecated_fields': common.get_json_data_deprecated_fields(json_data, schema_obj),
-        'json_data': json_data,  # Pass data so we can display values that need little processing
-        'common_error_types': []
-    })
-
-    return {
-        'context': context,
-        'cell_source_map': cell_source_map,
-    }
 
 
 def stats(request):
