+++ cove_360/views.py
@@ -4,32 +4,13 @@ import logging
 from django.shortcuts import render
 from django.utils.translation import ugettext_lazy as _
 
-from . lib.threesixtygiving import get_grants_aggregates, run_additional_checks
 from . lib.schema import Schema360
-from cove.lib.tools import datetime_or_date
+from . lib.threesixtygiving import common_checks_360
 from cove.lib.converters import convert_spreadsheet, convert_json
 from cove.lib.exceptions import CoveInputDataError, cove_web_input_error
-from cove.views import explore_data_context, common_checks_context
+from cove.views import explore_data_context
 
 logger = logging.getLogger(__name__)
-
-
-def common_checks_360(context, upload_dir, json_data, schema_obj):
-    schema_name = schema_obj.release_pkg_schema_name
-    checkers = {'date-time': (datetime_or_date, ValueError)}
-    common_checks = common_checks_context(upload_dir, json_data, schema_obj, schema_name, context, extra_checkers=checkers)
-    cell_source_map = common_checks['cell_source_map']
-    additional_checks = run_additional_checks(json_data, cell_source_map)
-
-    context.update(common_checks['context'])
-    context.update({
-        'grants_aggregates': get_grants_aggregates(json_data),
-        'additional_checks': additional_checks,
-        'additional_checks_count': len(additional_checks) + (1 if context['data_only'] else 0),
-        'common_error_types': ['uri', 'date-time', 'required', 'enum', 'integer', 'string']
-    })
-
-    return context
 
 
 @cove_web_input_error
