+++ cove_ocds/lib/ocds.py
@@ -1,6 +1,7 @@
 import collections
 
 import cove.lib.tools as tools
+from cove.lib.common import common_checks_context, get_additional_codelist_values
 
 
 @tools.ignore_errors
@@ -335,6 +336,33 @@ def get_releases_aggregates(json_data):
 
         contracts_without_awards=contracts_without_awards,
     )
+
+
+def common_checks_ocds(context, upload_dir, json_data, schema_obj, api=False):
+    schema_name = schema_obj.release_pkg_schema_name
+    if 'records' in json_data:
+        schema_name = schema_obj.record_pkg_schema_name
+    common_checks = common_checks_context(upload_dir, json_data, schema_obj, schema_name, context,
+                                          fields_regex=True, api=api)
+    validation_errors = common_checks['context']['validation_errors']
+
+    context.update(common_checks['context'])
+
+    if schema_name == 'record-package-schema.json':
+        context['records_aggregates'] = get_records_aggregates(json_data, ignore_errors=bool(validation_errors))
+        context['schema_url'] = schema_obj.record_pkg_schema_url
+    else:
+        additional_codelist_values = get_additional_codelist_values(schema_obj, schema_obj.codelists, json_data)
+        closed_codelist_values = {key: value for key, value in additional_codelist_values.items() if not value['isopen']}
+        open_codelist_values = {key: value for key, value in additional_codelist_values.items() if value['isopen']}
+
+        context.update({
+            'releases_aggregates': get_releases_aggregates(json_data, ignore_errors=bool(validation_errors)),
+            'additional_closed_codelist_values': closed_codelist_values,
+            'additional_open_codelist_values': open_codelist_values
+        })
+
+    return context
 
 
 @tools.ignore_errors
